implement prototype
