+++ Key2Joy.Gui/MainForm.resx
